Add pytest plugin to provide common Dallinger experiment test fixtures. Add new tet fixtures for running browser based tests against a debug instance. Adds an option to debug mode to not open browsers. Add new browser based tests for Bartlett demo. Refs #2029
