+++ dallinger/command_line.py
@@ -411,10 +411,17 @@ def get_summary(app):
 @click.option(
     "--proxy", default=None, help="Alternate port when opening browser windows"
 )
+@click.option(
+    "--no-browsers",
+    is_flag=True,
+    flag_value=True,
+    default=False,
+    help="Skip opening browsers",
+)
 @require_exp_directory
-def debug(verbose, bot, proxy, exp_config=None):
+def debug(verbose, bot, proxy, no_browsers=False, exp_config=None):
     """Run the experiment locally."""
-    debugger = DebugDeployment(Output(), verbose, bot, proxy, exp_config)
+    debugger = DebugDeployment(Output(), verbose, bot, proxy, exp_config, no_browsers)
     log(header, chevrons=False)
     debugger.run()
