+++ demos/tests/conftest.py
@@ -1,0 +1,3 @@
+"""Configure tests to use `pytest_dallinger` plugin"""
+
+pytest_plugins = ["pytest_dallinger"]
